Use a size_t for component sizes.
I guess we'll break whenever we need to support data with more
than 536,870,912 bytes per component.  I'm willing to take that
risk.

+++ Quantize.h
@@ -529,7 +529,7 @@ template <typename T, typename U>
 static bool BinningQuantize(LargeRAWFile& InputData,
                             const std::string& strTargetFilename,
                             uint64_t iSize,
-                            uint64_t& iComponentSize,
+                            size_t& iComponentSize,
                             Histogram1DDataBlock* Histogram1D=0)
 {
   MESSAGE("Attempting to recover integer values by binning the data.");
